--- docs/danyindex.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="danystyle.css" />
-    <script src="danyscript.js" defer></script>
-    <title>Document</title>
-  </head>
-  <body>
-    <button id="danyBtn"><div class="lds-heart"><div></button>
-    <h4 class="hidden" id="danyh4">DANY LOVES YOU</h4>
-    <h4 id="newh4" class="fade">Thanks for this class Alrik!</h4>
-  </body>
-</html>
