--- docs/danyscript.js
@@ -1,6 +1,0 @@
-let danyBtn = document.getElementById("danyBtn");
-let danyh4 = document.getElementById("danyh4");
-
-danyBtn.addEventListener("click", () => {
-  danyh4.classList.toggle("hidden");
-});
