--- docs/danystyle.css
@@ -1,119 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Aboreto&display=swap");
-
-#danyBtn {
-  font-family: "Chakra Petch", sans-serif;
-  background-color: black;
-  color: white;
-  border: 5px solid rgb(151, 3, 3);
-  cursor: pointer;
-  border-radius: 50%;
-}
-
-#danyBtn:hover {
-  filter: grayscale(1);
-}
-
-#danyh4 {
-  font-family: "Chakra Petch", sans-serif;
-  transition: 5s;
-  margin: 5px;
-  padding: 0;
-  text-decoration: overline rgb(151, 3, 3);
-  font-size: 35px;
-}
-
-.hidden {
-  display: none;
-}
-
-.fadeInText {
-  animation: fadeIn 5s;
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.lds-heart {
-  display: inline-block;
-  position: relative;
-  width: 80px;
-  height: 80px;
-  transform: rotate(45deg);
-  transform-origin: 40px 40px;
-}
-.lds-heart div {
-  top: 32px;
-  left: 32px;
-  position: absolute;
-  width: 32px;
-  height: 32px;
-  background: rgb(203, 18, 18);
-  animation: lds-heart 1.2s infinite cubic-bezier(0.215, 0.61, 0.355, 1);
-}
-.lds-heart div:after,
-.lds-heart div:before {
-  content: " ";
-  position: absolute;
-  display: block;
-  width: 32px;
-  height: 32px;
-  background: rgb(203, 18, 18);
-}
-.lds-heart div:before {
-  left: -24px;
-  border-radius: 50% 0 0 50%;
-}
-.lds-heart div:after {
-  top: -24px;
-  border-radius: 50% 50% 0 0;
-}
-@keyframes lds-heart {
-  0% {
-    transform: scale(0.95);
-  }
-  5% {
-    transform: scale(1.1);
-  }
-  39% {
-    transform: scale(0.85);
-  }
-  45% {
-    transform: scale(1);
-  }
-  60% {
-    transform: scale(0.95);
-  }
-  100% {
-    transform: scale(0.9);
-  }
-}
-
-#newh4 {
-  font-size: 40px;
-}
-
-#newh4:hover {
-  background-color: orange;
-  border: 5px red dotted;
-  transform: ease-in 3s;
-  width: 30%;
-}
-
-.fade {
-  animation: fadeIn 5s;
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
