+++ docs/index.html
@@ -6,11 +6,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>GRIT SQUARE</title>
     <link rel="stylesheet" href="style.css" />
+    <script src="script.js" type="module"></script>
   </head>
   <body oncontextmenu="return false;">
-   <!--  <input type="text" id="inName" />-->
-    <button id="btn">submit name</button>    <a href="./pages/AboutUs/aboutUs.html">About Us</a>
-<input type="color" name="" id="">
+    <!--  <input type="text" id="inName" />-->
+    <button id="btn">submit name</button> <a href="./pages/AboutUs/aboutUs.html">About Us</a>
+    <input type="color" name="" id="" />
     <div id="content"></div>
     <!-- <p class="bubble speech" style="top:20vw; left:30vh">sdfsdf sdf sd sdfdx df a ad asd dsdsHello world</p>
     <p class="bubble thought" >
@@ -18,7 +19,6 @@
       debitis.
     </p> 
     -->
-    <script src="script2.js" type="module"></script>
     <!-- <script src="script.js"></script> -->
   </body>
 </html>
