+++ docs/script.js
@@ -1,87 +1,196 @@
-import { initializeApp } from "https://www.gstatic.com/firebasejs/9.16.0/firebase-app.js";
-//import { getAnalytics } from "https://www.gstatic.com/firebasejs/9.16.0/firebase-analytics.js";
-import {
-  getDatabase,
-  ref,
-  set,
-  onValue,
-  remove,
-  push,
-} from "https://www.gstatic.com/firebasejs/9.16.0/firebase-database.js";
+import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.16.0/firebase-app.js';
+import { getDatabase, ref, set, onValue, remove, push, onChildAdded, onChildRemoved } from 'https://www.gstatic.com/firebasejs/9.16.0/firebase-database.js'; // <---
 
 // TODO: Add SDKs for Firebase products that you want to use
+
 // https://firebase.google.com/docs/web/setup#available-libraries
 
 // Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
+
 const firebaseConfig = {
-  databaseURL:
-    "https://onlinedatabase-2cd29-default-rtdb.europe-west1.firebasedatabase.app",
-  apiKey: "AIzaSyBuwc5HIp8-VaojCoNH904rph-YQwv9ugc",
-  authDomain: "onlinedatabase-2cd29.firebaseapp.com",
-  projectId: "onlinedatabase-2cd29",
-  storageBucket: "onlinedatabase-2cd29.appspot.com",
-  messagingSenderId: "513260891542",
-  appId: "1:513260891542:web:a3af24974e5dbd521381b7",
-  measurementId: "G-1QXQ3M2ZE2",
+  databaseURL: 'https://demo1-3c759-default-rtdb.europe-west1.firebasedatabase.app/', // <----
+  apiKey: 'AIzaSyDkRmoWLJ0eEXZKYEeUNwRF8V6X0oHwBi0',
+  authDomain: 'demo1-3c759.firebaseapp.com',
+  projectId: 'demo1-3c759',
+  storageBucket: 'demo1-3c759.appspot.com',
+  messagingSenderId: '632318648569',
+  appId: '1:632318648569:web:60f9813437fa829e598228',
 };
 
 // Initialize Firebase
-const app = initializeApp(firebaseConfig);
-//const analytics = getAnalytics(app);
-const database = getDatabase();
 
-// skriva
-function writeUserData(message) {
-  let adressRef = ref(database, "Eleonora/barn1");
-  set(adressRef, {
-    message: message,
-    username: "EL",
-    lista: [888, 2, 3, 4],
+const app = initializeApp(firebaseConfig);
+const db = getDatabase(app); // <-----
+
+console.log(db);
+
+//document.getElementById("btn").addEventListener(
+document.getElementById('content').addEventListener('click', (event) => {
+  var promptMessage = prompt('Write you message:', 'message');
+  if (promptMessage == null) return;
+
+  var e = event;
+  console.log(e);
+  var x = Math.round((event.clientX / window.innerWidth) * 100);
+  var y = Math.round((event.clientY / window.innerHeight) * 100);
+
+  console.log(window.innerWidth, window.innerHeight);
+  console.log(x, y);
+  //let inputName = document.getElementById("inName").value;
+  //console.log(inputName);
+  /*
+    set(ref(db, inputName), {
+        dateOfCretion: new Date().toString("yyyy-MM-dd hh:mm:ss"),
+        message: promptMessage,
+        x: x,
+        y: y,
+    });*/
+  push(ref(db, '/'), {
+    username: 'Alrik',
+    dateOfCretion: new Date().toString('yyyy-MM-dd hh:mm:ss'),
+    message: promptMessage,
+    x: x,
+    y: y,
+  });
+});
+
+function writeUserData() {
+  let inputName = document.getElementById('inName').value;
+  console.log(inputName);
+  set(ref(db, inputName), {
+    message: 'hello',
+    dateOfCretion: new Date(),
   });
 }
+//writeUserData()
 
-writeUserData("SUPP DOG!!!");
-
-// läsa en specifik onValue direkt + när den ändrar value
-const urlRef = ref(database, "Eleonora/barn1");
-onValue(urlRef, (snapshot) => {
-  const data = snapshot.val();
-  console.log(data.lista);
-  document.body.innerHTML = data.lista[1];
-});
-
-const urlRootRef = ref(database, "/Eleonora");
-//läsa childs i root.
-console.warn("LOOP and children!!");
-onValue(
-  urlRootRef,
+/* onValue(
+  ref(db, 'Alrik'),
   (snapshot) => {
-    snapshot.forEach((childSnapshot) => {
-      const childKey = childSnapshot.key;
-      const childData = childSnapshot.val();
-      console.log(childKey);
-      console.log(childData.message);
-    });
-  } /* ,
-    {
-        onlyOnce: true,  // engång
-    } */
-);
+    const data = snapshot.val();
+    //alert(data.message)
+    //document.body.innerHTML= data.message;
+  },
+  { onlyOnce: true }
+); */
 
-//radera stuff.
-const deleteRef = ref(database, "/filip/username");
-remove(deleteRef).then(() => {
-  console.log("location removed");
+onChildAdded(ref(db, '/'), (data) => {
+  let d = data.val();
+
+  if (d.message.length < 5) document.getElementById('content').insertAdjacentHTML('beforeend', `<p class="bubble" id="${data.key}" style="left:${d.x}vw; top:${d.y}vh">${d.message}</p>`);
+  else document.getElementById('content').insertAdjacentHTML('beforeend', `<p class="bubble speech" id="${data.key}" style="left:${d.x}vw; top:${d.y}vh">${d.message}</p>`);
+
+  document.getElementById(data.key).addEventListener('contextmenu', (event) => event.preventDefault());
+
+  let bubbleID = document.querySelector(`#${data.key}`);
+
+  document.getElementById(data.key).addEventListener('mouseup', (event) => {
+    if (event.button == 2) {
+      alert('Delete message?');
+      bubbleID.remove();
+      remove(ref(db, bubbleID.id));
+    }
+  });
+
+  /*document.getElementById(
+    "content"
+  ).innerHTML += `<p class="bubble speech" style="left:${d.x}vw; top:${d.y}vh">${d.message}</p>`;*/
+  console.log(d);
 });
 
-// kolla om det supportas i olika webb läsare
-if (typeof Storage !== "undefined") {
-  // läsa från user pref
-  console.log(localStorage.getItem("name"));
-  console.log(localStorage.getItem("lastLoggin"));
+onChildRemoved(ref(db, '/'), (data) => {
+  console.log(document.querySelector(`#${data.key}`).remove());
+  console.log(document.getElementById(data.key).remove());
+});
 
-  //skriva i user pref
-  //localStorage.setItem("name", "Alrik");
-  // localStorage.setItem("lastLoggin","today!!");
+/*
+remove(ref(db , 'henrik') ).then(() => {
+    console.log("alrik removed");
+});*/
+
+//db.ref("-Users/-KUanJA9egwmPsJCxXpv").update({ displayName: "New trainer" });
+
+//update( ref( db, alrik), () =>{ message: "New trainer" } );
+
+document.getElementById('btn').addEventListener('click', get);
+
+// Alriks databas
+const BASE_URL = 'https://demo1-3c759-default-rtdb.europe-west1.firebasedatabase.app/.json';
+
+/* PUT , läggaer till skriver över */
+async function putMessage() {
+  let messageObject = {
+    text: 'Hello world put',
+    time: new Date(),
+  };
+  console.log(JSON.stringify(messageObject));
+  const requestOptions = {
+    method: 'PUT',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(messageObject),
+  };
+
+  let response = await fetch(BASE_URL, requestOptions);
+  let data = await response.json();
+  console.log(data);
+}
+
+/* PATCH , skriver/uppdaterar men skriver ej över */
+async function patchMessage() {
+  let messageObject = { text: 'Hello world put', time: new Date() };
+
+  const requestOptions = {
+    method: 'PATCH',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(messageObject),
+  };
+
+  let response = await fetch(BASE_URL, requestOptions);
+  let data = await response.json();
+  console.log(data);
+}
+async function patchMessage2() {
+  let messageObject = { firstName: 'Alrik', lastName: 'HE' };
+
+  const requestOptions = {
+    method: 'PATCH',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(messageObject),
+  };
+
+  let response = await fetch('https://demo1-3c759-default-rtdb.europe-west1.firebasedatabase.app/-NoCf2s5SInOIFR3X0VX/.json', requestOptions);
+  let data = await response.json();
+  console.log(data);
+}
+
+/* POST */
+async function postMessage() {
+  let messageObject = { text: 'Hello world', time: new Date() };
+
+  const requestOptions = {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(messageObject),
+  };
+
+  let response = await fetch(BASE_URL, requestOptions);
+  let data = await response.json();
+  console.log(data);
+}
+
+/* GET */
+async function getMessages() {
+  let response = await fetch(BASE_URL);
+  let data = await response.json();
+  console.log(data);
+}
+
+/* DELETE */
+async function deleteMessage() {
+  const requestOptions = {
+    method: 'DELETE',
+  };
+  let response = await fetch(BASE_URL, requestOptions);
+  let data = await response.json();
+  console.log(data);
 }
